+++ R/get_global_range.R
@@ -15,7 +15,7 @@
 #'
 #' @examples
 #' \dontrun{
-#' get_global_range(result = result_attr)
+#' get_global_range(result = result_rand)
 #' }
 #'
 #' @aliases get_global_range